TOML

TOML.jl is a Julia standard library for parsing and writing TOML v1.0 files.

Parsing TOML data

julia> using TOML

julia> data = """
           [database]
           server = "192.168.1.1"
           ports = [ 8001, 8001, 8002 ]
       """;

julia> TOML.parsestring(data)
Dict{String,Any} with 1 entry:
  "database" => Dict{String,Any}("server"=>"192.168.1.1","ports"=>Any[8001, 800…

To parse a file, use TOML.parsefile. If the file has a syntax error, an exception is thrown:

julia> using TOML

julia> TOML.parsestring("""
           value = 0.0.0
       """)
ERROR: TOML Parser error:
none:1:16 error: failed to parse value
      value = 0.0.0
                 ^
[...]

There are other versions of the parse functions (TOML.tryparsestring and [TOML.tryparsefile]) that instead of throwing exceptions on parser error returns a TOML.ParserError with information:

julia> using TOML

julia> err = TOML.tryparsestring("""
           value = 0.0.0
       """);

julia> err.type
ErrGenericValueError::ErrorType = 14

julia> err.line
1

julia> err.column
16

Exporting data to TOML file

The TOML.print function is used to print (or serialize) data into TOML format.

julia> using TOML

julia> fname = tempname();

julia> data = Dict(
          "names" => ["Julia", "Julio"],
          "age" => [10, 20],
       );

julia> open(fname, "w") do io
           TOML.print(io, data)
       end;

julia> print(read(fname, String))
names = ["Julia", "Julio"]
age = [10, 20]

References

TOML.parsestring — Function

parsestring(str::AbstractString)
parsestring(p::Parser, str::AbstractString)

Parses a string str and returns the resulting table (dictionary). Returns a ParserError upon failure.

See also TOML.tryparsestring

TOML.parsefile — Function

parsefile(f::AbstractString)
parsefile(p::Parser, f::AbstractString)

Parses a file f and returns the resulting table (dictionary). Throws a ParserError upon failure.

See also TOML.tryparsefile

TOML.tryparsestring — Function

tryparsestring(str::AbstractString)
tryparsestring(p::Parser, str::AbstractString)

Parses a string str and returns the resulting table (dictionary). Returns a ParserError upon failure.

See also TOML.parsestring

TOML.tryparsefile — Function

tryparsefile(f::AbstractString)
tryparsefile(p::Parser, f::AbstractString)

Parses a file f and returns the resulting table (dictionary). Returns a ParserError upon failure.

See also TOML.parsefile

TOML.print — Function

print(io::IO, data::AbstractDict)

Writes data into TOML syntax to the stream io.

TOML.Parser — Type

Parser()

Constructor for a TOML Parser. Note that in most cases one does not need to explicitly create a Parser but instead one directly use use TOML.parsefile or parsestring. Using an explicit parser will however reuse some internal data structures which can be beneficial for performance if a larger number of small files are parsed.

TOML.ParserError — Type

ParserError

Type that is returned from tryparsestring and tryparsefile when parsing fails. It contains (among others) the following fields:

pos, the position in the string when the error happened

table, the result that so far was successfully parsed

type, an error type, different for different types of errors